+++ CHANGELOG.md
@@ -7,6 +7,9 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+- Publish deleted proposals
+
 ## [0.3.1] - 2023-12-04
 
 ### Fixed